move the kchaodb logic from blockstore to dbmanager
