+++ src/main/java/org/tron/core/db/BlockStore.java
@@ -15,20 +15,13 @@
 
 package org.tron.core.db;
 
-import java.util.ArrayList;
-import java.util.stream.Collectors;
-import javafx.util.Pair;
-import org.apache.commons.lang3.ArrayUtils;
 import org.joda.time.DateTime;
 import org.joda.time.format.DateTimeFormat;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.tron.common.storage.leveldb.LevelDbDataSourceImpl;
-import org.tron.common.utils.ByteArray;
 import org.tron.core.Sha256Hash;
 import org.tron.core.capsule.BlockCapsule;
 import org.tron.core.capsule.TransactionCapsule;
-import org.tron.core.config.args.Args;
 import org.tron.core.exception.ValidateException;
 
 public class BlockStore extends TronDatabase {
@@ -37,18 +30,14 @@ public class BlockStore extends TronDatabase {
   //private LevelDbDataSourceImpl blockDbDataSource;
   //private LevelDbDataSourceImpl unSpendCache;
 
-  private LevelDbDataSourceImpl numHashCache;
+  //private LevelDbDataSourceImpl numHashCache;
 
-  private KhaosDatabase khaosDb;
+  //private KhaosDatabase khaosDb;
 
   private BlockCapsule head;
 
   private BlockStore(String dbName) {
     super(dbName);
-    numHashCache = new LevelDbDataSourceImpl(
-        Args.getInstance().getOutputDirectory(), dbName + "_NUM_HASH");
-    numHashCache.initDB();
-    khaosDb = new KhaosDatabase(dbName + "_KDB");
   }
 
   private static BlockStore instance;
@@ -67,9 +56,6 @@ public class BlockStore extends TronDatabase {
     return instance;
   }
 
-  public void initHeadBlock(Sha256Hash id) {
-    head = getBlockByHash(id);
-  }
 
   /**
    * to do.
@@ -85,37 +71,6 @@ public class BlockStore extends TronDatabase {
     return head == null ? 0 : head.getNum();
   }
 
-  /**
-   * Get the block id from the number.
-   */
-  public Sha256Hash getBlockIdByNum(long num) {
-    byte[] hash = numHashCache.getData(ByteArray.fromLong(num));
-    return ArrayUtils.isNotEmpty(hash) ? Sha256Hash.wrap(hash) : Sha256Hash.ZERO_HASH;
-  }
-
-  /**
-   * Get number of block by the block id.
-   */
-  public long getBlockNumById(Sha256Hash hash) {
-    if (khaosDb.containBlock(hash)) {
-      return khaosDb.getBlock(hash).getNum();
-    }
-
-    //TODO: optimize here
-    byte[] blockByte = dbSource.getData(hash.getBytes());
-    return ArrayUtils.isNotEmpty(blockByte) ? new BlockCapsule(blockByte).getNum() : 0;
-  }
-
-  /**
-   * Get the fork branch.
-   */
-  public ArrayList<Sha256Hash> getBlockChainHashesOnFork(Sha256Hash forkBlockHash) {
-    Pair<ArrayList<BlockCapsule>, ArrayList<BlockCapsule>> branch =
-        khaosDb.getBranch(head.getBlockId(), forkBlockHash);
-    return branch.getValue().stream()
-        .map(blockCapsule -> blockCapsule.getBlockId())
-        .collect(Collectors.toCollection(ArrayList::new));
-  }
 
   public DateTime getHeadBlockTime() {
     return head == null ? getGenesisTime() : new DateTime(head.getTimeStamp());
@@ -130,22 +85,6 @@ public class BlockStore extends TronDatabase {
     return DateTime.parse("20180101", DateTimeFormat.forPattern("yyyyMMdd"));
   }
 
-  /**
-   * judge id.
-   *
-   * @param blockHash blockHash
-   */
-  public boolean containBlock(Sha256Hash blockHash) {
-    //TODO: check it from levelDB
-    return khaosDb.containBlock(blockHash) || dbSource.getData(blockHash.getBytes()) != null;
-  }
-
-  /**
-   * judge has blocks.
-   */
-  public boolean hasBlocks() {
-    return dbSource.allKeys().size() > 0 || khaosDb.hasData();
-  }
 
   /**
    * push transaction into db.
@@ -159,32 +98,6 @@ public class BlockStore extends TronDatabase {
     return true;
   }
 
-  /**
-   * find a block packed data by id.
-   */
-  public byte[] findBlockByHash(Sha256Hash hash) {
-    return khaosDb.containBlock(hash) ? khaosDb.getBlock(hash).getData()
-        : dbSource.getData(hash.getBytes());
-  }
-
-  /**
-   * Get a BlockCapsule by id.
-   */
-  public BlockCapsule getBlockByHash(Sha256Hash hash) {
-    return khaosDb.containBlock(hash) ? khaosDb.getBlock(hash)
-        : new BlockCapsule(dbSource.getData(hash.getBytes()));
-  }
-
-  /**
-   * Delete a block.
-   */
-  public void deleteBlock(Sha256Hash blockHash) {
-    BlockCapsule block = getBlockByHash(blockHash);
-    khaosDb.removeBlk(blockHash);
-    dbSource.deleteData(blockHash.getBytes());
-    numHashCache.deleteData(ByteArray.fromLong(block.getNum()));
-    head = khaosDb.getHead();
-  }
 
   public void getUnspend(byte[] key) {
   }
