+++ src/main/java/org/tron/core/net/node/NodeDelegateImpl.java
@@ -64,7 +64,7 @@ public class NodeDelegateImpl implements NodeDelegate {
       //todo: find a block we all know between the summary and my db.
       Collections.reverse(blockChainSummary);
       for (Sha256Hash hash : blockChainSummary) {
-        if (getBlockStoreDb().containBlock(hash)) {
+        if (dbManager.containBlock(hash)) {
           lastKnownBlkHash = hash;
           break;
         }
@@ -75,10 +75,10 @@ public class NodeDelegateImpl implements NodeDelegate {
       }
     }
 
-    for (long num = getBlockStoreDb().getBlockNumById(lastKnownBlkHash);
+    for (long num = dbManager.getBlockNumById(lastKnownBlkHash);
         num <= getBlockStoreDb().getHeadBlockNum(); ++num) {
       if (num > 0) {
-        retBlockHashes.add(getBlockStoreDb().getBlockIdByNum(num));
+        retBlockHashes.add(dbManager.getBlockIdByNum(num));
       }
     }
     return retBlockHashes;
@@ -96,12 +96,12 @@ public class NodeDelegateImpl implements NodeDelegate {
 
     if (refPoint != Sha256Hash.ZERO_HASH) {
       //todo: get db's head num to check local db's block status.
-      if (getBlockStoreDb().containBlock(refPoint)) {
-        highBlkNum = getBlockStoreDb().getBlockNumById(refPoint);
+      if (dbManager.containBlock(refPoint)) {
+        highBlkNum = dbManager.getBlockNumById(refPoint);
         highNoForkBlkNum = highBlkNum;
       } else {
-        forkList = getBlockStoreDb().getBlockChainHashesOnFork(refPoint);
-        highNoForkBlkNum = getBlockStoreDb().getBlockNumById(forkList.get(forkList.size() - 1));
+        forkList = dbManager.getBlockChainHashesOnFork(refPoint);
+        highNoForkBlkNum = dbManager.getBlockNumById(forkList.get(forkList.size() - 1));
         forkList.remove(forkList.get(forkList.size() - 1));
       }
 
@@ -116,7 +116,7 @@ public class NodeDelegateImpl implements NodeDelegate {
     long realHighBlkNum = highBlkNum + num;
     do {
       if (lowBlkNum <= highNoForkBlkNum) {
-        retSummary.add(getBlockStoreDb().getBlockIdByNum(lowBlkNum));
+        retSummary.add(dbManager.getBlockIdByNum(lowBlkNum));
       } else {
         retSummary.add(forkList.get((int) (lowBlkNum - highNoForkBlkNum - 1)));
       }
@@ -130,7 +130,7 @@ public class NodeDelegateImpl implements NodeDelegate {
   public Message getData(Sha256Hash hash, MessageTypes type) {
     switch (type) {
       case BLOCK:
-        return new BlockMessage(getBlockStoreDb().findBlockByHash(hash));
+        return new BlockMessage(dbManager.findBlockByHash(hash));
       case TRX:
         return new TransactionMessage(
             dbManager.getTransactionStore().findTransactionByHash(hash.getBytes()));
@@ -163,7 +163,7 @@ public class NodeDelegateImpl implements NodeDelegate {
   @Override
   public boolean contain(Sha256Hash hash, MessageTypes type) {
     if (type.equals(MessageTypes.BLOCK)) {
-      return getBlockStoreDb().containBlock(hash);
+      return dbManager.containBlock(hash);
     } else if (type.equals(MessageTypes.TRX)) {
       //TODO: check it
       return false;
